Add support for 24x24 tile set.
I compared the Drill2x code to the Miss Driller code to find all pieces
that had to be changed. However, I replaced hardcoded coordinates and sizes
with constants. The main constant is TILE_SIZE, which is set by the
Makefile; the other constants are defined in the new header "geometry.h".